requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-pptx==0.6.23
pytesseract==0.3.10
Pillow==10.2.0

# LLM Integration - Gemini
google-generativeai==0.3.2

# Background Tasks
celery==5.3.6
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
hypothesis==6.98.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
jsonschema==4.20.0
